Remove old tof.frames module
