+++ .buildconfig/ci-linux.yml
@@ -38,3 +38,4 @@ dependencies:
   - sphinx-book-theme==0.3.3
   - sphinx-copybutton==0.5.2
   - sphinxcontrib-bibtex==2.6.1
+  - sciline==24.01.1
